requirements.txt
spotipy
Flask
gunicorn
python-dotenv